requirements.txt
django
markdown-it-py
